commit e77274decf6924a2cfe244c1cf59dd95ec170417
    14856 cmn_err(9F) missing vdev_err()
commit bc0ee17c150fbf29e52c0ff365163e4e7b1c2f0a
    15226 scf_tmpl_pg_name(3SCF) missing const on some functions
commit c743cdfdcf2e342944e95263a263b01bdc7f20d9
    15323 cond_init(3C) is confused about NULL being a pointer
commit 810da10bc091b2f30a41135f4466e937ae77b518
    16152 ddi_add_event_handler(9F) has wrong argument type
commit 5c2b9f0df71da9360641a9feba64a92b312a005a
    16741 ddi_intr_add_handler(9F) should probably document ddi_intr_handler_t
commit 5d99edaf58f86c30708894b62fe4ccbadeb12c8f
    16186 ndd is obsolete for TCP/IP too
commit f82c9dfee6174146c0855dd18558eb197e9527a4
    16783 dladm(8) has incorrect fields for "show-iptun"
commit 8a721ada1217887e4e82a903719b503ce6d5ab6a
    16609 taskq(9f): unrelated parameter tp
commit 97d970a8ffe4a62af961a3d47bee8d73cc9dc7e0
    15542 wcswidth(3c): example formatting error
commit c3fc68cb9bee4a79c80730cf9a0c7772b68a80bc
    16527 cxgbe: want support for CPL_TX_PKT_XT
commit 8cd244d2cd9581fbeca0eb25a292ddf847423af9
    16811 unix: 'repl_pfn' may be used uninitialized
commit 01d4d8e79088c1c743ce26509694aae9b9e08696
    16812 fsck_pcfs: need to support 4kn sector size

Conflicts:
    manifest
    usr/src/uts/intel/mac/Makefile

+++ usr/src/uts/common/sys/mac_provider.h
@@ -675,7 +675,7 @@ typedef struct mac_ether_offload_info {
 	size_t		meoi_len;	/* Total message length */
 	uint8_t		meoi_l2hlen;	/* How long is the Ethernet header? */
 	uint16_t	meoi_l3proto;	/* What's the Ethertype */
-	uint8_t		meoi_l3hlen;	/* How long is the header? */
+	uint16_t	meoi_l3hlen;	/* How long is the header? */
 	uint8_t		meoi_l4proto;	/* What is the payload type? */
 	uint8_t		meoi_l4hlen;	/* How long is the L4 header */
 } mac_ether_offload_info_t;
